Fix the binary I/O to restore the flashrom support for BusPirate v4
flashrom (opensource flashing tool) starts with a binary mode and then
wants to return to the interactive mode to get the version and stuff.
The method to reset - works only for BPv3; USB of BPv4 does not like it.
So: for BPv4, use another method to get back to the interactive mode.

I have reviewed this commit that has been done by @denis2342
at his Bus Pirate private repository and believe it is OK.

Co-authored-by: Denis Ahrens <denis@h3q.com>
Signed-off-by: Mike Banon <mikebdp2@gmail.com>