Restore the SPI speed table compatibility with the flashrom FOSS tool
A change about 8 years ago made the SPI speed table incompatible
with the flashrom (FOSS flashing tool): flashrom thinks that it sets 8MHz
SPI speed when in reality it sets only 2.6MHz. So: lets change this table
back to the state it was, to make it (again) compatible with the flashrom tool.

I have reviewed this commit that has been done by @denis2342 at
his Bus Pirate private repository, and a minor fix I believe it is OK

Co-authored-by: Denis Ahrens <denis@h3q.com>
Signed-off-by: Mike Banon <mikebdp2@gmail.com>